requirements.txt
streamlit==1.51.0
tensorflow-cpu==2.20.0
ultralytics==8.3.227
numpy==2.2.6
pillow==12.0.0
requests==2.32.5
opencv-python==4.12.0.88
torch==2.9.0
torchvision==0.24.0
